requirements-old.txt
# Framework Web et Serveur
fastapi>=0.111.0
uvicorn[standard]==0.27.1
gunicorn==23.0.0

# Base de données et Cache
redis>=4.5.0
motor>=3.0.0,<4.0.0
pymongo==4.6.3
numpy>=1.24.0

# IA et Traitement des données
openai==1.67.0
numpy>=1.21.0
sentence-transformers>=3.1.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Analyse de graphe
networkx>=3.2.1
python-louvain>=0.16
community>=1.0.0

# Utilitaires et Sécurité
pydantic>=1.10.0,<3.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.18
slowapi==0.1.8
bech32==1.2.0
ecdsa==0.19.1

# Tests
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Dépendances pour l'évaluation
scikit-learn>=1.0.2
matplotlib>=3.5.1
seaborn>=0.11.2
pymongo>=4.1.1
pytest>=7.0.0
numpy>=1.21.5

# RAG et Monitoring
prometheus-client>=0.16.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
python-dotenv>=0.19.0
redis>=4.5.0
numpy>=1.24.0
aiohttp>=3.8.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
openai>=1.0.0

# Nouvelles dépendances
aiohttp==3.9.3
beautifulsoup4==4.12.3
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
selenium==4.18.1
webdriver-manager==4.0.1

# Optimisations
faiss-cpu==1.10.0
sentence-transformers==4.0.2
zlib==1.2.13
